Merging this PR as it was open for quite a while now. The mebx files produced with this code have been verified by multiple MPEG members.

+++ IsoLib/libisomediafile/src/CompatibleSchemeTypeAtom.c
@@ -1,26 +1,15 @@
-/*
- * This software module was originally developed by InterDigital, Inc.
- * in the course of development of MPEG-4.
- * This software module is an implementation of a part of one or
- * more MPEG-4 tools as specified by MPEG-4.
- * ISO/IEC gives users of MPEG-4 free license to this
- * software module or modifications thereof for use in hardware
- * or software products claiming conformance to MPEG-4 only for evaluation and testing purposes.
- * Those intending to use this software module in hardware or software
- * products are advised that its use may infringe existing patents.
- * The original developer of this software module and his/her company,
- * the subsequent editors and their companies, and ISO/IEC have no
- * liability for use of this software module or modifications thereof
- * in an implementation.
- *
- * Copyright is not released for non MPEG-4 conforming
- * products. InterDigital, Inc. retains full right to use the code for its own
- * purpose, assign or donate the code to a third party and to
- * inhibit third parties from using the code for non
- * MPEG-4 conforming products.
- *
- * This copyright notice must be included in all copies or
- * derivative works.
+/* This software module was originally developed by InterDigital, Inc. in the course of development
+ * of MPEG-4. This software module is an implementation of a part of one or more MPEG-4 tools as
+ * specified by MPEG-4. ISO/IEC gives users of MPEG-4 free license to this software module or
+ * modifications thereof for use in hardware or software products claiming conformance to MPEG-4
+ * only for evaluation and testing purposes. Those intending to use this software module in hardware
+ * or software products are advised that its use may infringe existing patents. The original
+ * developer of this software module and his/her company, the subsequent editors and their
+ * companies, and ISO/IEC have no liability for use of this software module or modifications thereof
+ * in an implementation. Copyright is not released for non MPEG-4 conforming products. InterDigital,
+ * Inc. retains full right to use the code for its own purpose, assign or donate the code to a third
+ * party and to inhibit third parties from using the code for non MPEG-4 conforming products. This
+ * copyright notice must be included in all copies or derivative works.
  */
 
 /**
@@ -71,7 +60,6 @@ static MP4Err serialize(struct MP4Atom *s, char *buffer)
   assert(self->bytesWritten == self->size);
 bail:
   TEST_RETURN(err);
-
   return err;
 }
 
@@ -94,7 +82,6 @@ static MP4Err calculateSize(struct MP4Atom *s)
 
 bail:
   TEST_RETURN(err);
-
   return err;
 }
 
@@ -114,7 +101,7 @@ static MP4Err createFromInputStream(MP4AtomPtr s, MP4AtomPtr proto, MP4InputStre
   GET32(scheme_type);
   GET32(scheme_version);
 
-  bytesToRead = self->size - self->bytesRead;
+  bytesToRead = (long)self->size - self->bytesRead;
   if(bytesToRead < 0) BAILWITHERROR(MP4BadDataErr);
 
   if(bytesToRead > 0)
@@ -148,7 +135,6 @@ static MP4Err createFromInputStream(MP4AtomPtr s, MP4AtomPtr proto, MP4InputStre
 
 bail:
   TEST_RETURN(err);
-
   return err;
 }
 
@@ -175,6 +161,5 @@ MP4Err MP4CreateCompatibleSchemeTypeAtom(MP4CompatibleSchemeTypeAtomPtr *outAtom
 
 bail:
   TEST_RETURN(err);
-
   return err;
 }
